// frontend/src/types/ColorsEnum.ts
enum ColorsEnum {
  RED = "rgb(229,19,23)",
  BORDO = "rgb(118,16,55)",
  PINK = "rgb(208,69,98)",
  VIOLET = "rgb(77,71,115)",
  GRAY = "rgb(193,195,196)",
  GRAYDARK = "rgba(255, 255, 255, 0.12)",
  BGLIGHT = "rgb(58,65,73)",
  BGLIGHTER = "rgb(65,73,80)",
  BGDARK = "rgb(47,53,58)",
  WHITE = "white",
  BLACK = "black",
  BLUE = "rgb(61,152,247)",
  OLIVE = "#3D9970",
  ORANGE = "#ff8a65",
  ERROR = "#ff6961", // RED1 = red[500],
  GREEN = "green",
  GOLD = "#E8BF8B",
}

export const GraphColors = [
  "#E5BC76",
  "#8ABABE",
  "#7f78ac",
  "#E45558",
  "#16B5A7",
  "#63BE9C",
  "green",
  "green",
  "green",
  "green",
  "green",
  "green",
  "green",
  "green",
  "green",
];

export default ColorsEnum;